requirements.txt
django
celery[redis]
python-decouple